Redo boilerplate (type checking disabled for now)
# Conflicts:
#	bitwarden_keyring/__init__.py
