+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,11 @@
+Closes #<ticket number>
+
+### Checklist:
+<!-- In case of doubt, we're here to help. CONTRIBUTING.rst might help too -->
+- [ ] Tests
+  - [ ] (not applicable?)
+- [ ] Documentation (README)
+  - [ ] (not applicable?)
+
+<!-- We hope your contributing experience was great so far. If you would like to
+provide some feedback, please feel free to do so! -->
