+++ .github/renovate.json5
@@ -1,0 +1,17 @@
+{
+    extends: ["config:recommended", ":enablePreCommit"],
+    schedule: ["before 4am on Saturday"],
+    labels: ["PR type: dependencies 🤖"],
+    lockFileMaintenance: {
+        enabled: true,
+        automerge: true,
+    },
+    packageRules: [
+        {
+            groupName: "all dependencies",
+            groupSlug: "all",
+            matchPackageNames: ["*"],
+        },
+    ],
+    automerge: true,
+}
