+++ .gitignore
@@ -3,3 +3,4 @@ build
 dist
 .coverage
 htmlcov
+.venv
