--- .travis.yml
@@ -1,14 +1,0 @@
-# Config file for automatic testing at travis-ci.org
-language: python
-python: 3.6
-
-cache: pip
-
-install:
-  - pip install -r requirements.txt codecov
-
-script:
-  - pytest
-
-after_success:
-- codecov
