--- setup.cfg
@@ -1,55 +1,0 @@
-[metadata]
-name = bitwarden-keyring
-description = Keyring backend reading password data from Bitwarden
-version = 0.3.1
-author = Joachim Jablon
-author_email = ewjoachim@gmail.com
-url = https://github.com/ewjoachim/bitwarden-keyring
-long_description = file: README.md
-long_description_content_type = text/markdown
-keywords = bitwarden keyring password
-license = Apache Software License
-classifiers =
-    Development Status :: 4 - Beta
-    Intended Audience :: Developers
-    License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 3
-
-[options]
-zip_safe = True
-include_package_data = True
-packages = find:
-install_requires =
-    keyring
-
-[options.entry_points]
-keyring.backends =
-    bitwarden = bitwarden_keyring
-
-[options.extras_require]
-
-dev =
-    flake8
-    isort
-    black
-    pytest
-    pytest-mock
-    pytest-cov
-
-
-[flake8]
-max-line-length = 88
-
-[isort]
-multi_line_output=3
-include_trailing_comma=True
-force_grid_wrap=0
-use_parentheses=True
-line_length=88
-not_skip = __init__.py
-
-[bdist_wheel]
-universal = 1
-
-[tool:pytest]
-addopts = --cov-report term-missing --cov-branch --cov-report html --cov-report term --cov=bitwarden_keyring -vv
