--- setup.py
@@ -1,3 +1,0 @@
-from setuptools import setup
-
-setup()
